+++ Sources/App/Core/SiteURL.swift
@@ -126,7 +126,7 @@ enum SiteURL: Resourceable {
     case stylesheets(String)
     case supporters
     case tryInPlayground
-    case validateSpiManifest
+    case validateSPIManifest
 
     var path: String {
         switch self {
@@ -216,7 +216,7 @@ enum SiteURL: Resourceable {
             case .tryInPlayground:
                 return "try-in-a-playground"
 
-            case .validateSpiManifest:
+            case .validateSPIManifest:
                 return "validate-spi-manifest"
         }
     }
@@ -224,7 +224,7 @@ enum SiteURL: Resourceable {
     var pathComponents: [PathComponent] {
         switch self {
             case .addAPackage, .buildMonitor, .faq, .home, .packageCollections, .privacy, .rssPackages,
-                    .rssReleases, .search, .siteMap, .supporters, .tryInPlayground, .validateSpiManifest:
+                    .rssReleases, .search, .siteMap, .supporters, .tryInPlayground, .validateSPIManifest:
                 return [.init(stringLiteral: path)]
 
             case let .api(next):
